apexstudio/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.17
slowapi==0.1.9
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
python-dotenv==1.0.1
